requirements.txt
alpaca-py==0.8.2
black==23.7.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
exceptiongroup==1.1.3
fonttools==4.39.4
hurst==0.0.5
ibapi==9.81.1.post1
idna==3.4
importlib-resources==5.12.0
iniconfig==2.0.0
kiwisolver==1.4.4
matplotlib==3.7.1
msgpack==1.0.5
mypy-extensions==1.0.0
numpy==1.24.3
packaging==23.1
pandas==1.5.3
pathspec==0.11.1
patsy==0.5.3
Pillow==9.5.0
platformdirs==3.9.1
pluggy==1.2.0
polygon-api-client==1.13.0
pydantic==1.10.12
pyparsing==3.0.9
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
scipy==1.11.2
six==1.16.0
sseclient-py==1.7.2
statsmodels==0.14.0
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.18
websockets==10.4
zipp==3.15.0
